TRUNK-925: Session endpoint  should respect the name template
