+++ omod-2.7/src/test/resources/sessionControllerTestDataset.xml
@@ -1,0 +1,39 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!--
+    This Source Code Form is subject to the terms of the Mozilla Public License,
+    v. 2.0. If a copy of the MPL was not distributed with this file, You can
+    obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+    the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+    Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+    graphic logo is a trademark of OpenMRS Inc.
+-->
+<dataset>
+	<person person_id="5501" gender="M" dead="false" creator="1" date_created="2014-01-09 00:00:00.0" voided="false" uuid="1716432a-4bf9-436c-baad-f03236e960e1"/>
+	<person_name person_name_id="5501" preferred="true" person_id="5501" prefix="Mr." given_name="Moses" middle_name="Tusha" family_name="Mujuzi" family_name2="Jr" creator="1" date_created="2014-01-09 00:00:00.0" voided="false" uuid="54fdf8e8-1a80-45cf-a3fd-4fabd1b8ffab"/>
+	<users user_id="5501" person_id="5501" system_id="4-2" username="mujuzi" password="4a1750c8607d0fa237de36c6305715c223415189" salt="c788c6ad82a157b712392ca695dfcf2eed193d7f" creator="1" date_created="2014-01-09 00:00:00.0" retired="false" uuid="06d05314-e132-11de-babe-001e37123456"/>
+
+	<global_property property="layout.name.format" property_value="customXmlTemplate" uuid="1716432a-1a80-11de-iuj7-1b2e21e4df9c"/>
+    <global_property property="layout.name.template" property_value="&lt;org.openmrs.layout.name.NameTemplate&gt;
+    &lt;codeName&gt;customXmlTemplate&lt;/codeName&gt;
+    &lt;displayName&gt;Custom Name Format&lt;/displayName&gt;
+    &lt;nameMappings class=&quot;properties&quot;&gt;
+      &lt;property name=&quot;prefix&quot; value=&quot;PersonName.prefix&quot;/&gt;
+      &lt;property name=&quot;givenName&quot; value=&quot;PersonName.givenName&quot;/&gt;
+      &lt;property name=&quot;familyName&quot; value=&quot;PersonName.familyName&quot;/&gt;
+    &lt;/nameMappings&gt;
+    &lt;sizeMappings class=&quot;properties&quot;&gt;
+      &lt;property name=&quot;prefix&quot; value=&quot;5&quot;/&gt;
+      &lt;property name=&quot;givenName&quot; value=&quot;30&quot;/&gt;
+	  &lt;property name=&quot;familyName&quot; value=&quot;25&quot;/&gt;
+	&lt;/sizeMappings&gt;
+	&lt;lineByLineFormat&gt;
+	  &lt;string&gt;prefix&lt;/string&gt;
+	  &lt;string&gt;givenName&lt;/string&gt;
+	  &lt;string&gt;familyName&lt;/string&gt;
+	&lt;/lineByLineFormat&gt;
+	&lt;requiredElements&gt;
+      &lt;string&gt;givenName&lt;/string&gt;
+      &lt;string&gt;familyName&lt;/string&gt;
+    &lt;/requiredElements&gt;
+  &lt;/org.openmrs.layout.name.NameTemplate&gt;" uuid="87uy5tee3-4aa7-5674-bad4-1b2e21e4df9c"/>
+</dataset>
