+++ pom.xml
@@ -428,6 +428,7 @@
 				<module>omod-2.2</module>
 				<module>omod-2.3</module>
 				<module>omod-2.4</module>
+				<module>omod-2.7</module>
 				<module>omod</module>
 				<module>integration-tests</module>
 			</modules>
@@ -501,6 +502,7 @@
 				<module>omod-2.2</module>
 				<module>omod-2.3</module>
 				<module>omod-2.4</module>
+				<module>omod-2.7</module>
 				<module>omod</module>
 				<module>integration-tests</module>
 			</modules>
